\name{print.ctsmr}
\alias{print.ctsmr}
\title{Print the CTSMR class}
\usage{
\method{print}{ctsmr}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
  \item{x}{an object of class "ctsmr".}

  \item{digits}{the number of significant digits to use
  when printing.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
Print the CTSMR class
}

