\name{print.summary.ctsmr}
\alias{print.summary.ctsmr}
\title{Print the summary of a CTSMR class}
\usage{
\method{print}{summary.ctsmr}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
  \item{x}{an object of class "summary.ctsmr".}

  \item{digits}{the number of significant digits to use
  when printing.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
Print the summary of a CTSMR class
}

